Update with --dry-run flag

+++ remove-duplicate.sh
@@ -1,19 +1,47 @@
 #!/bin/sh
 
-CWD=$(pwd)
-#filename=$1
+dryRun=false
 
+# Parse options
+for opt in $*; do
+
+	# Check for option
+	if [[ $opt == -* ]]; then
+		# Shift away flags
+		shift
+
+		# Check for dry run
+		if [ $opt == "--dry-run" ]; then
+			dryRun=true
+		fi
+
+		if [ $opt == "--help" ]; then
+			echo "usage: remove-duplicate [--dry-run] [--help] file1 [file2 [...]]"
+			echo
+			echo "       --dry-run  List all files that would be removed, do not remove anything"
+			echo "       --help     Show this help message"
+			exit;
+		fi
+	fi
+done
+
+CWD=$(pwd)
+
+# Do for each file
 for filename in $*
 do
+	# Dirname
 	dir=$(dirname "$CWD/$filename")
+	# Filename
 	filename=$(basename "$filename")
 
 	# Check that file exists
 	if [ ! -e "$dir/$filename" ]; then
 		echo "File $dir/$filename does not exist"
-		exit 1;
+		continue
 	fi
 
+	# Calculate the hash for file
 	hash=$(shasum "$dir/$filename" | sed -E 's/([a-z0-9]+).*$/\1/')
 
 	# A string containing all sha1 hashes
@@ -21,20 +49,18 @@ do
 	# Remove current file from list
 	allHashes=$(echo $allHashes | sed -E "s/[a-z0-9]+ $filename//")
 
-	#echo $hash
-	#echo $allHashes
-
 	# Grep for hash
 	grepResult=$(echo "$allHashes" | grep "$hash")
 
 	if [ "$grepResult" != "" ]; then
-		#echo "This is a duplicate removing"
-		rm "$dir/$filename"
-	#else
-		#echo "This is not a duplicate"
+		
+		# If dry run, do not remove anything
+		if [ $dryRun ]; then
+			echo "$dir/$filename"
+		else
+			# Remove file
+			rm "$dir/$filename"
+		fi
 	fi
 
 done
-
-
-#echo $shasums
